requirements.txt
Django==5.2.8
djangorestframework==3.16.1
uvicorn==0.38.0